Move JITDebugInfoRegistry to its own file
